Access key

+++ src/controllers/data.ts
@@ -1,6 +1,7 @@
 import {NextFunction, Request, Response, Router} from 'express';
 
 import {LogEntry} from '../models/LogEntry';
+import {getSecrets} from "../helpers/config-loader";
 
 const router = Router();
 
@@ -11,6 +12,12 @@ router.get('/', (req: Request, res: Response, next: NextFunction) => {
 });
 
 router.post('/', (req: Request, res: Response, next: NextFunction) => {
+	const config = getSecrets();
+	if (req.header("Authorization") !== `Bearer ${config.accessKey}`) {
+		res.status(403).end();
+		return
+	}
+
 	const body = <Buffer> req.body;
 	const parsedBody = <any[]> JSON.parse(body.toString());
 	LogEntry.destroy({truncate: true}).then(() => {
